requirements.txt
streamlit==1.29.0
python-dotenv==1.0.0
google-generativeai==0.3.1
wordcloud==1.9.2
pandas==2.1.4
matplotlib==3.8.2
plotly==5.18.0